backend/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
pillow
pydicom
numpy
python-multipart
pytest
pytest-asyncio
httpx 